Beta day1

+++ Git_program/STM32-v2.1/Schdule/device.h
@@ -36,7 +36,7 @@ typedef struct{
     uint16 pyh_address;									//物理地址
     uint8  data;												//数据
 	  uint16 ab_slot_num;									//绝对时隙号
-		uint8  keep_alive;									//alive标志位 
+		uint32  keep_alive;									//alive标志位 
 		
 }EndPointDeviceStruct;
